Dev Estoque, Precificacao e inicio consolidado

+++ GestaoComercio/GestaoComercio.Application/Services/FornecedorService.cs
@@ -3,9 +3,11 @@ using GestaoComercio.Application.Models.Fornecedor.Commands;
 using GestaoComercio.Application.Models.Pedido.Commands;
 using GestaoComercio.Domain.Entities;
 using GestaoComercio.Domain.Interfaces;
+using GestaoComercio.Domain.Utils;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -15,12 +17,14 @@ namespace GestaoComercio.Application.Services
     {
 
         private readonly IGenericRepository<Fornecedor> _fornecedorRepository;
+        private readonly IGenericRepository<Pedido> _pedidoRepository;
         private readonly IMapper _mapper;
 
-        public FornecedorService(IGenericRepository<Fornecedor> fornecedorRepository, IMapper mapper)
+        public FornecedorService(IGenericRepository<Fornecedor> fornecedorRepository, IMapper mapper, IGenericRepository<Pedido> pedidoRepository)
         {
             _fornecedorRepository = fornecedorRepository;
             _mapper = mapper;
+            _pedidoRepository = pedidoRepository;
         }
         public async Task<FornecedorDTO> InserirFornecedor(PostFornecedorCommand request)
         {
@@ -35,6 +39,12 @@ namespace GestaoComercio.Application.Services
 
         public async Task<FornecedorDTO> DeletarFornecedor(string cnpj)
         {
+            var pedidos = _pedidoRepository.GetAsync().Result.Where(x => x.CodigoFornecedorProduto == cnpj);
+            if (pedidos.Count() > 0)
+            {
+                throw new MyExceptionApi("Não é possivel deletar o Fornecedor, pois há pedidos atrelados a ele!", HttpStatusCode.BadRequest);
+            }
+
             var fornecedor = _fornecedorRepository.Get(x => x.Cnpj == cnpj.ToString());
             return _mapper.Map<FornecedorDTO>(await _fornecedorRepository.RemoveAsync(fornecedor));
         }
